# be/contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Private Clinic API
  version: 0.17.0
  description: Public contract for the private clinic management system.
servers:
  - url: http://localhost:5000
paths:
...
  /api/v1/admin/catalog/reference-data:
    get:
      operationId: getCatalogReferenceData
      summary: Get manageable branches and catalog references
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Catalog reference data scoped to the actor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogReferenceResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/admin/catalog/organization:
    get:
      operationId: getOrganizationCatalog
      summary: List branches, specialties and categories for global catalog management
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Organization catalog including inactive records.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationCatalogResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/admin/catalog/branches:
    post:
      operationId: createOrganizationBranch
      summary: Create a branch
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateOrganizationBranchRequest' }
      responses:
        '201':
          description: Created branch.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrganizationBranchResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/admin/catalog/branches/{branchId}:
    put:
      operationId: updateOrganizationBranch
      summary: Update branch policies or activity
      security:
        - bearerAuth: []
      parameters:
        - name: branchId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateOrganizationBranchRequest' }
      responses:
        '200':
          description: Updated branch.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrganizationBranchResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }
  /api/v1/admin/catalog/specialties:
    post:
      operationId: createOrganizationSpecialty
      summary: Create specialty
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateOrganizationSpecialtyRequest' }
      responses:
        '201':
          description: Created specialty.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrganizationSpecialtyResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/admin/catalog/specialties/{specialtyId}:
    put:
      operationId: updateOrganizationSpecialty
      summary: Update specialty
      security:
        - bearerAuth: []
      parameters:
        - name: specialtyId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateOrganizationSpecialtyRequest' }
      responses:
        '200':
          description: Updated specialty.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrganizationSpecialtyResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }
  /api/v1/admin/catalog/categories:
    post:
      operationId: createOrganizationCategory
      summary: Create service category
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateOrganizationCategoryRequest' }
      responses:
        '201':
          description: Created category.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrganizationCategoryResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/admin/catalog/categories/{categoryId}:
    put:
      operationId: updateOrganizationCategory
      summary: Update service category
      security:
        - bearerAuth: []
      parameters:
        - name: categoryId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateOrganizationCategoryRequest' }
      responses:
        '200':
          description: Updated category.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrganizationCategoryResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }
  /api/v1/admin/catalog/rooms:
    get:
      operationId: listCatalogRooms
      summary: List rooms in a manageable branch
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BranchPublicIdQuery'
      responses:
        '200':
          description: Rooms in the selected branch.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogRoomListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      operationId: createCatalogRoom
      summary: Create a room in a manageable branch
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCatalogRoomRequest'
      responses:
        '201':
          description: Room created.
          headers:
            ETag:
              $ref: '#/components/headers/CatalogEtag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogRoomResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/admin/catalog/rooms/{roomId}:
    put:
      operationId: updateCatalogRoom
      summary: Update a room with optimistic concurrency
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/CatalogIfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCatalogRoomRequest'
      responses:
        '200':
          description: Room updated.
          headers:
            ETag:
              $ref: '#/components/headers/CatalogEtag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogRoomResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '428':
          $ref: '#/components/responses/PreconditionRequired'
  /api/v1/admin/catalog/services:
    get:
      operationId: listCatalogServices
      summary: List organization services and branch price history
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BranchPublicIdQuery'
      responses:
        '200':
          description: Organization services with selected-branch prices.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogServiceListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: createCatalogService
      summary: Create an organization-wide service
      description: Requires a global MASTER_DATA_MANAGE assignment. Branch managers cannot create organization catalog definitions.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BranchPublicIdQuery'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCatalogServiceRequest'
      responses:
        '201':
          description: Organization service created.
          headers:
            ETag:
              $ref: '#/components/headers/CatalogEtag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogServiceResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/admin/catalog/services/{serviceId}:
    put:
      operationId: updateCatalogService
      summary: Update an organization-wide service with optimistic concurrency
      security:
        - bearerAuth: []
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/BranchPublicIdQuery'
        - $ref: '#/components/parameters/CatalogIfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCatalogServiceRequest'
      responses:
        '200':
          description: Organization service updated.
          headers:
            ETag:
              $ref: '#/components/headers/CatalogEtag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogServiceResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '428':
          $ref: '#/components/responses/PreconditionRequired'
  /api/v1/admin/catalog/services/{serviceId}/prices:
    post:
      operationId: setCatalogBranchPrice
      summary: Schedule a branch price and availability change
      security:
        - bearerAuth: []
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetCatalogBranchPriceRequest'
      responses:
        '201':
          description: Branch price period scheduled without overlap.
          headers:
            ETag:
              $ref: '#/components/headers/CatalogEtag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogServiceResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  parameters:
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Strong ETag containing the Base64 employee rowversion and optional doctor rowversion.
      schema:
        type: string
        pattern: '^"[A-Za-z0-9+/]{11}=(:[A-Za-z0-9+/]{11}=)?"$'
    CatalogIfMatch:
      name: If-Match
      in: header
      required: true
      description: Strong ETag containing the Base64 rowversion of a catalog resource.
      schema:
        type: string
        pattern: '^"[A-Za-z0-9+/]{11}="$'
...
    BranchPublicIdQuery:
      name: branchPublicId
      in: query
      required: true
      schema:
        type: string
        pattern: '^[0-9A-Fa-f]{8}-(?:[0-9A-Fa-f]{4}-){3}[0-9A-Fa-f]{12}$'
...
  headers:
    CatalogEtag:
      description: Strong ETag containing the Base64 SQL rowversion.
      schema:
        type: string
        pattern: '^"[A-Za-z0-9+/]{11}="$'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Unauthorized:
      description: Authentication data is missing, invalid, expired, reused, or revoked.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Forbidden:
      description: The account is unavailable for login.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    NotFound:
      description: The requested resource does not exist in the caller's scope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Conflict:
      description: The operation conflicts with uniqueness, concurrency, or account safety rules.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    PreconditionRequired:
      description: The If-Match precondition is required for this update.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
...
  schemas:
...
    ResponseMeta:
      type: object
      additionalProperties: true
    RequestId:
      type: string
      format: uuid
...
    CatalogRoomType:
      type: string
      enum: [CONSULTATION, PROCEDURE, LAB, IMAGING, PHARMACY, OTHER]
...
    CreateOrganizationBranchRequest:
      allOf:
        - $ref: '#/components/schemas/OrganizationBranchFields'
        - type: object
          required: [code]
          properties:
            code: { type: string, pattern: '^[A-Z0-9_-]{2,20}$' }
    UpdateOrganizationBranchRequest:
      allOf:
        - $ref: '#/components/schemas/OrganizationBranchFields'
        - type: object
          required: [isActive]
          properties:
            isActive: { type: boolean }
...
    CreateOrganizationSpecialtyRequest:
      type: object
      additionalProperties: false
      required: [code, name, description]
      properties:
        code: { type: string, pattern: '^[A-Z0-9_-]{2,30}$' }
        name: { type: string, minLength: 2, maxLength: 150 }
        description: { type: [string, 'null'], maxLength: 1000 }
    UpdateOrganizationSpecialtyRequest:
      type: object
      additionalProperties: false
      required: [name, description, isActive]
      properties:
        name: { type: string, minLength: 2, maxLength: 150 }
        description: { type: [string, 'null'], maxLength: 1000 }
        isActive: { type: boolean }
    CreateOrganizationCategoryRequest:
      type: object
      additionalProperties: false
      required: [code, name, displayOrder]
      properties:
        code: { type: string, pattern: '^[A-Z0-9_-]{2,30}$' }
        name: { type: string, minLength: 2, maxLength: 150 }
        displayOrder: { type: integer, minimum: 0, maximum: 1000000 }
    UpdateOrganizationCategoryRequest:
      type: object
      additionalProperties: false
      required: [name, displayOrder, isActive]
      properties:
        name: { type: string, minLength: 2, maxLength: 150 }
        displayOrder: { type: integer, minimum: 0, maximum: 1000000 }
        isActive: { type: boolean }
    OrganizationCatalogResponse:
      type: object
      additionalProperties: false
      required: [data, meta, requestId]
      properties:
        data:
          type: object
          additionalProperties: false
          required: [branches, specialties, categories]
          properties:
            branches:
              type: array
              items: { $ref: '#/components/schemas/OrganizationBranch' }
            specialties:
              type: array
              items: { $ref: '#/components/schemas/OrganizationSpecialty' }
            categories:
              type: array
              items: { $ref: '#/components/schemas/OrganizationCategory' }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
        requestId: { $ref: '#/components/schemas/RequestId' }
    OrganizationBranchResponse:
      type: object
      additionalProperties: false
      required: [data, meta, requestId]
      properties:
        data: { $ref: '#/components/schemas/OrganizationBranch' }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
        requestId: { $ref: '#/components/schemas/RequestId' }
    OrganizationSpecialtyResponse:
      type: object
      additionalProperties: false
      required: [data, meta, requestId]
      properties:
        data: { $ref: '#/components/schemas/OrganizationSpecialty' }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
        requestId: { $ref: '#/components/schemas/RequestId' }
    OrganizationCategoryResponse:
      type: object
      additionalProperties: false
      required: [data, meta, requestId]
      properties:
        data: { $ref: '#/components/schemas/OrganizationCategory' }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
        requestId: { $ref: '#/components/schemas/RequestId' }
...
    CatalogReferenceResponse:
      type: object
      additionalProperties: false
      required: [data, meta, requestId]
      properties:
        data:
          $ref: '#/components/schemas/CatalogReferenceData'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
        requestId:
          $ref: '#/components/schemas/RequestId'
...
    CatalogRoomResponse:
      type: object
      additionalProperties: false
      required: [data, meta, requestId]
      properties:
        data:
          $ref: '#/components/schemas/CatalogRoom'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
        requestId:
          $ref: '#/components/schemas/RequestId'
    CatalogRoomListResponse:
      type: object
      additionalProperties: false
      required: [data, meta, requestId]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/CatalogRoom'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
        requestId:
          $ref: '#/components/schemas/RequestId'
...
    CreateCatalogRoomRequest:
      type: object
      additionalProperties: false
      required: [branchPublicId, code, name, type, capacity]
      properties:
        branchPublicId:
          type: string
          format: uuid
        code:
          type: string
          minLength: 1
          maxLength: 30
        name:
          type: string
          minLength: 2
          maxLength: 150
        type:
          $ref: '#/components/schemas/CatalogRoomType'
        floorNo:
          type: integer
          minimum: -5
          maximum: 200
        capacity:
          type: integer
          minimum: 1
          maximum: 500
...